Move validator to generic
