+++ golem-worker-service-base/src/service/http/mod.rs
@@ -1,0 +1,1 @@
+mod http_api_definition_validator;
